Added to services

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/JdbcUserDao.java
@@ -61,9 +61,9 @@ public class JdbcUserDao implements UserDao {
         // create user
         String sql = "INSERT INTO users (username, password_hash) VALUES (?, ?) RETURNING user_id";
         String password_hash = new BCryptPasswordEncoder().encode(password);
-        Integer newUserId;
+        Long newUserId;
         try {
-            newUserId = jdbcTemplate.queryForObject(sql, Integer.class, username, password_hash);
+            newUserId = jdbcTemplate.queryForObject(sql, Long.class, username, password_hash);
         } catch (DataAccessException e) {
             return false;
                 }
